prepare to introduce cdi injection
