+++ nosql/common/src/main/java/org/wildfly/nosql/common/ConnectionServiceAccess.java
@@ -1,0 +1,46 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2016, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.wildfly.nosql.common;
+
+import org.jboss.as.server.CurrentServiceContainer;
+import org.jboss.msc.service.ServiceName;
+import org.wildfly.nosql.common.spi.NoSQLConnection;
+
+/**
+ * ConnectionServiceAccess
+ *
+ * @author Scott Marlow
+ */
+public class ConnectionServiceAccess {
+
+    private static final ServiceName SERVICENAME = ServiceName.JBOSS.append("NoSQLClientConnectionService");
+
+    public static ServiceName serviceName(String id) {
+        return SERVICENAME.append(id);
+    }
+
+    public static NoSQLConnection connection(String id) {
+        return (NoSQLConnection) CurrentServiceContainer.getServiceContainer().getRequiredService(SERVICENAME.append(id));
+    }
+
+}
